+++ CHANGELOG.md
@@ -5,6 +5,9 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## Added
+- **Experimental**, gateway: add UUID support for VPN gateway connections and tunnels
+
 ## [8.5.0]
 
 ## Added